pyproject.toml
[project]
name = "hopsworks-featurestore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.9.0",
    "hopsworks>=4.2.0",
    "mlflow>=2.21.3",
    "pandas>=2.1.4",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.1.0",
    "setuptools>=78.1.0",
    "sweetviz>=2.3.1",
]
